pyproject.toml
[project]
name = "telegramify-markdown"
version = "1.0.0"
description = "Convert Markdown to Telegram plain text + MessageEntity pairs"
authors = [
    { name = "sudoskys", email = "coldlando@hotmail.com" },
]
dependencies = [
    "pyromark>=0.7.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
mermaid = [
    "Pillow>=10.4.0",
    "aiohttp>=3.10.11",
]
tests = [
    "pyTelegramBotAPI>=4.22.0",
    "python-dotenv>=1.0.1",
]

[project.urls]
repository = "https://github.com/sudoskys/telegramify-markdown"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.scripts]
test = "python -m unittest discover -s ./tests -p test_*.py"
